dice roll command moved to cog
